nixos: refactor system config deployment
We replace the single hard-coded nix configuration with both a
hcloud-optimized default and a user-supplied configuration.  This user
configuration mirrors the one usually in /etc/nixos/, with the intent
of making customized nixos hosts simple to set up. The user can always
forgo this mechanism and only include the relevant /etc/nixos/hcloud/
config.