archlinux: add support for dynamic initialization of VMs
Some data is only known once VMs are instantiated; this includes
root's SSH Keys, hostname, and IP. We download and aggregate this data
from Hetzner's endpoints on initial boot of the VM, and then trigger
services that implement the following features based on this data.

This provides some semblance of feature parity to the standard hetzner
VMs, and exposes a unified view of the data to the user. This enables,
for example, using the --user-data-from-file and --ssh-key flags in
the hcloud CLI for archlinux VMs.

Note that the NixOS support for this is still incomplete.